Add Columns class.
